requirements.txt
torch>=2.0
torchvision
transformers
albumentations
opencv-python
Pillow
tensorboard
tqdm
